Handwritten Digit Recognition using CNN
Though the goal is to create a model which can recognize the digits, we can extend it for letters and then a person’s handwriting. 
Through this work, we aim to learn and practically apply the concepts of Convolutional Neural Networks. 
We are going to add voice while displaying the output.